etl/requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
curl_cffi==0.13.0
frozendict==2.4.6
idna==3.10
multitasking==0.0.12
numpy==2.3.2
pandas==2.3.2
peewee==3.18.2
platformdirs==4.4.0
protobuf==6.32.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.65
